Added some functions for the classes, changed score points
Changed score points to 10 per question.
Added a timeout for applying the class of colour to correct or incorrect questions.
The green or red class will be removed after 800ms.
added a function for the incrementScore variable to add your score to the inner text of the scoreText element.
Not currently working, need to find a fix